- Refractored most part of the controllers, pending LogController. - Corrected bugs

+++ PushApi/Controllers/PreferenceController.php
@@ -7,7 +7,6 @@ use \PushApi\Controllers\Controller;
 use \PushApi\Models\User;
 use \PushApi\Models\Theme;
 use \PushApi\Models\Preference;
-use \Illuminate\Database\QueryException;
 use \Illuminate\Database\Eloquent\ModelNotFoundException;
 
 /**
@@ -18,21 +17,6 @@ use \Illuminate\Database\Eloquent\ModelNotFoundException;
 class PreferenceController extends Controller
 {
     /**
-     * Retrives all preferences of a given user
-     * @param [int] $idUser User identification
-     */
-    public function getPreferences($idUser)
-    {
-        try {
-            $preferences = User::findOrFail($idUser)->preferences()->orderBy('id', 'asc')->get();
-        } catch (ModelNotFoundException $e) {
-            throw new PushApiException(PushApiException::NOT_FOUND);
-        }
-
-        $this->send($preferences->toArray());
-    }
-
-    /**
      * Sets user preference to a given theme, if the preference has
      * been done before, it only displays the information of the preference
      * else, creates the preference and displays the resulting information
@@ -41,49 +25,64 @@ class PreferenceController extends Controller
      */
     public function setPreference($idUser, $idTheme)
     {
-        try {
-            $option = (int) $this->slim->request->post('option');
+        $option = (int) $this->slim->request->post('option');
 
-            if (!isset($option)) {
-                throw new PushApiException(PushApiException::NO_DATA);
-            }
-
-            if ($update['option'] > Preference::ALL_RANGES) {
-                throw new PushApiException(PushApiException::INVALID_OPTION);
-            }
-
-            $preference = User::find($idUser)->preferences()->where('theme_id', $idTheme)->first();
-            if (!isset($preference) || empty($preference)) {
-                if (!empty(User::find($idUser)->toArray()) && !empty(Theme::find($idTheme)->toArray())) {
-                    $preference = new Preference;
-                    $preference->user_id = $idUser;
-                    $preference->theme_id = $idTheme;
-                    $preference->option = $option;
-                    $preference->save();
-                }
-            }
-        } catch (QueryException $e) {
-            throw new PushApiException(PushApiException::NOT_FOUND);
-        } catch (\Exception $e) {
-            throw new PushApiException(PushApiException::INVALID_ACTION);
+        if (!isset($option)) {
+            throw new PushApiException(PushApiException::NO_DATA);
         }
-        $this->send($preference->toArray());
-    }
 
-    /**
-     * Retrives preference of a given theme
-     * @param [int] $idUser User identification
-     * @param [int] $idTheme Theme identification
-     */
-    public function getPreference($idUser, $idTheme)
-    {
+        if ($option > Preference::ALL_RANGES) {
+            throw new PushApiException(PushApiException::INVALID_OPTION);
+        }
+
+        // Checking if preference already exists
         try {
-            $preferences = User::findOrFail($idUser)->preferences()->where('theme_id', $idTheme)->orderBy('id', 'asc')->get();
+            $preference = User::findOrFail($idUser)->preferences()->where('theme_id', $idTheme)->first();
         } catch (ModelNotFoundException $e) {
             throw new PushApiException(PushApiException::NOT_FOUND);
         }
 
-        $this->send($preferences->toArray());
+        if (!isset($preference) || empty($preference)) {
+            if (!empty(Theme::find($idTheme))) {
+                $preference = new Preference;
+                $preference->user_id = $idUser;
+                $preference->theme_id = $idTheme;
+                $preference->option = $option;
+                $preference->save();
+            }
+        }
+
+        if (!empty($preference)) {
+            $this->send($preference->toArray());
+        } else {
+            $this->send(array());
+        }
+    }
+
+    /**
+     * Retrives all preferences of a given user or it also can check
+     * if user has set preferences from a theme (if he has set it, the
+     * preference is displayed)
+     * @param [int] $idUser User identification
+     * @param [int] $idTheme Theme identification
+     */
+    public function getPreference($idUser, $idTheme = false)
+    {
+        try {
+            if ($idTheme) {
+                $preferences = User::findOrFail($idUser)->preferences()->where('theme_id', $idTheme)->orderBy('id', 'asc')->get();
+            } else {
+                $preferences = User::findOrFail($idUser)->preferences()->orderBy('id', 'asc')->get();
+            }
+        } catch (ModelNotFoundException $e) {
+            throw new PushApiException(PushApiException::NOT_FOUND);
+        }
+
+        if (!empty($preferences)) {
+            $this->send($preferences->toArray());
+        } else {
+            $this->send(array());
+        }
     }
 
     /**
@@ -93,23 +92,21 @@ class PreferenceController extends Controller
      */
     public function updatePreference($idUser, $idTheme)
     {
+        $update = array();
+        $update['option'] = (int) $this->slim->request->post('option');
+
+        if (!isset($update['option'])) {
+            throw new PushApiException(PushApiException::NO_DATA);
+        }
+
+        if ($update['option'] > Preference::ALL_RANGES) {
+            throw new PushApiException(PushApiException::INVALID_OPTION);
+        }
+
         try {
-            $update = array();
-            $update['option'] = (int) $this->slim->request->post('option');
-
-            if (!isset($update['option'])) {
-                throw new PushApiException(PushApiException::NO_DATA);
-            }
-
-            if ($update['option'] > Preference::ALL_RANGES) {
-                throw new PushApiException(PushApiException::INVALID_OPTION);
-            }
-
             $user = Preference::where('theme_id', $idTheme)->update($update);
         } catch (ModelNotFoundException $e) {
             throw new PushApiException(PushApiException::NOT_FOUND);
-        } catch (\Exception $e) {
-            throw new PushApiException(PushApiException::INVALID_ACTION);
         }
         $this->send($this->boolinize($user));
     }
@@ -123,12 +120,15 @@ class PreferenceController extends Controller
     {
         try {
             $preference = User::findOrFail($idUser)->preferences()->where('theme_id', $idTheme)->first();
-            $preference->delete();
         } catch (ModelNotFoundException $e) {
             throw new PushApiException(PushApiException::NOT_FOUND);
-        } catch (\Exception $e) {
-            throw new PushApiException(PushApiException::INVALID_ACTION);
         }
-        $this->send($preference->toArray());
+
+        if (!empty($preference)) {
+            $preference->delete();
+            $this->send($preference->toArray());
+        } else {
+            throw new PushApiException(PushApiException::NOT_FOUND);
+        }
     }
 }
